+++ GUI/script/danbooru/tab.py
@@ -3,11 +3,16 @@ import typing as t
 
 from PySide6 import QtCore
 from PySide6.QtCore import Qt, Signal
+from PySide6.QtGui import QIcon
 from PySide6.QtWidgets import QCompleter, QFrame, QHBoxLayout, QVBoxLayout, QWidget
-from qfluentwidgets import Action, ComboBox, FluentIcon as FIF, FlowLayout, PushButton, RoundMenu, ScrollArea, SearchLineEdit, TransparentToolButton
+from qfluentwidgets import (
+    Action, ComboBox, EditableComboBox, FluentIcon as FIF, FlowLayout, PrimaryToolButton, ToolButton, 
+    RoundMenu, ScrollArea, SearchLineEdit, TeachingTipTailPosition, TransparentToolButton
+)
 from qfluentwidgets.components.widgets.line_edit import CompleterMenu
 
 from GUI.uic.qfluent import MonkeyPatch as FluentMonkeyPatch
+from GUI.uic.qfluent.components import CustomTeachingTip
 from utils.config.qc import danbooru_cfg
 from utils.script.image.danbooru.constants import DANBOORU_SORT_OPTIONS
 from utils.script.image.danbooru.models import DanbooruAutocompleteCandidate, DanbooruPost
@@ -27,14 +32,17 @@ class DanbooruTabWidget(QFrame):
     request_next_page = Signal()
     detail_opened = Signal(object)
     request_close = Signal()
+    request_extra_search = Signal(str)
 
     SORT_OPTIONS = list(DANBOORU_SORT_OPTIONS)
 
     def __init__(self, state: DanbooruTabState, parent=None):
         super().__init__(parent)
+        self.gui = parent.gui
         self.state = state
         self.card_metrics = DEFAULT_CARD_METRICS
         self.card_widgets: dict[str, DanbooruCardWidget] = {}
+        self._extra_tip = None
         self._setup_ui()
         self.selection_controller = DanbooruTabSelectionController(self)
         self.selection_controller.selection_count_changed.connect(self.selection_count_changed.emit)
@@ -59,6 +67,9 @@ class DanbooruTabWidget(QFrame):
         query_frame, query_group = self._create_group_frame("DanbooruSearchQueryGroup")
         self.query_frame = query_frame
         self.query_group = query_group
+        self.extraSearchBtn = TransparentToolButton(FIF.ADD, self)
+        self.extraSearchBtn.setVisible(False)
+        self.extraSearchBtn.clicked.connect(self._on_extra_search_clicked)
         self.search_edit = SearchLineEdit(self)
         self.search_edit.setClearButtonEnabled(True)
         self.search_edit.setPlaceholderText("such as: blue_archive")
@@ -66,10 +77,12 @@ class DanbooruTabWidget(QFrame):
         self.search_edit.returnPressed.connect(self._submit_search_from_keyboard)
         self.search_edit.searchSignal.connect(lambda text: self.request_search.emit(text))
         self.search_edit.searchButton.clicked.connect(self._submit_empty_search_if_needed)
+        self.search_edit.textChanged.connect(self._sync_extra_search_btn_visibility)
         self.set_search_menu()
         self.favorite_btn = TransparentToolButton(FIF.HEART, self)
         self.favorite_btn.setFixedSize(38, 38)
-        self.convert_btn = PushButton("to Tag", self)
+        self.convert_btn = ToolButton(QIcon(':/script/translate.svg'), self)
+        self.convert_btn.setIconSize(QtCore.QSize(24,24))
         self.convert_btn.setMinimumHeight(38)
         self.convert_btn.clicked.connect(self.request_conversion.emit)
         self.sort_box = ComboBox(self)
@@ -77,7 +90,8 @@ class DanbooruTabWidget(QFrame):
         for label, _ in self.SORT_OPTIONS:
             self.sort_box.addItem(label)
         self.sort_box.currentIndexChanged.connect(self._on_sort_changed)
-        query_group.addWidget(self.search_edit, 1)
+        query_group.addWidget(self.extraSearchBtn)
+        query_group.addWidget(self.search_edit)
         query_group.addWidget(self.favorite_btn)
         query_group.addWidget(self.convert_btn)
         query_group.addWidget(self.sort_box)
@@ -100,6 +114,9 @@ class DanbooruTabWidget(QFrame):
         self.scroll_area.verticalScrollBar().valueChanged.connect(self._on_scroll_changed)
         self.main_layout.addWidget(self.scroll_area, 1)
         self.refresh_from_state()
+
+    def _sync_extra_search_btn_visibility(self):
+        self.extraSearchBtn.setVisible(bool(self.search_edit.text().strip()))
 
     def _set_search_edit_value(self, value: str):
         self.search_edit.setText(value)
@@ -131,8 +148,8 @@ class DanbooruTabWidget(QFrame):
             submenu.setIcon(FIF.HEART)
             groups = visible_usage_groups(
                 build_tag_groups(
-                    sorted(danbooru_cfg.get_favorites()),
-                    danbooru_cfg.get_grouped_favorites()))
+                    sorted(danbooru_cfg.fav.get()),
+                    danbooru_cfg.fav.get_grouped()))
             actions = [Action(text=group.display,
                 triggered=lambda _=False, current=list(group.tags): _show_group_completer(current))
             for group in groups]
@@ -184,6 +201,32 @@ class DanbooruTabWidget(QFrame):
         self.state.sort_mode = value
         self.request_search.emit(self.search_edit.text())
 
+    def _on_extra_search_clicked(self):
+        if self._extra_tip is not None:
+            self._extra_tip.close()
+        extraCombo = EditableComboBox(self)
+        extraCombo.setMinimumWidth(120)
+        extraCombo.setPlaceholderText("e.g. score:>50")
+        extraCombo.addItems(danbooru_cfg.get_search_extra())
+        helpBtn = TransparentToolButton(FIF.HELP, self)
+        helpBtn.clicked.connect(lambda: self.gui.open_url_by_browser(
+            "https://www.yuque.com/baimusheng/programer/wl9c6nxxdvecm1tg"))
+        svBtn = PrimaryToolButton(FIF.ACCEPT_MEDIUM, self)
+        tip = CustomTeachingTip.create(
+            [extraCombo, svBtn, helpBtn],
+            target=self.extraSearchBtn, parent=self,
+            tailPosition=TeachingTipTailPosition.BOTTOM_LEFT,
+        )
+        self._extra_tip = tip
+        tip.destroyed.connect(lambda *_args: setattr(self, '_extra_tip', None))
+        def _apply():
+            value = extraCombo.currentText().strip()
+            if value:
+                danbooru_cfg.add_search_extra(value)
+                self.request_extra_search.emit(value)
+            tip.close()
+        svBtn.clicked.connect(_apply)
+
     def _on_scroll_changed(self, value: int):
         bar = self.scroll_area.verticalScrollBar()
         if bar.maximum() - value < 200 and not self.state.loading and self.state.has_more_results:
@@ -216,6 +259,7 @@ class DanbooruTabWidget(QFrame):
         self.search_edit.searchButton.setDisabled(loading)
         self.convert_btn.setDisabled(loading)
         self.sort_box.setDisabled(loading)
+        self.extraSearchBtn.setDisabled(loading)
 
     def clear_images_before_current_page(self):
         if not self.state.has_pages_before_current():